+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Enable markdown autolinks for `gemini://` URL scheme.
 - Don't pollute PostgreSQL log with `activitypub_object_pkey` constraint violation errors.
+- Improved error reporting during actor document parsing.
 
 ## [3.7.1] - 2024-10-17